chore(release): 📦  version packages (#228)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @llm-ui/json@0.10.0

### Minor Changes

- [#226](https://github.com/llm-ui-kit/llm-ui/pull/226)
[`0fdafd8`](https://github.com/llm-ui-kit/llm-ui/commit/0fdafd8d7575d225c7b03afdafc7d361c2eff11c)
Thanks [@richardgill](https://github.com/richardgill)! - -
jsonBlockExample params are now an object
    -   jsonBlockPrompt params `userOptions` -> `options`

### Patch Changes

-   Updated dependencies \[]:
    -   @llm-ui/react@0.10.0
    -   @llm-ui/shared@0.10.0

## @llm-ui/code@0.10.0

### Patch Changes

-   Updated dependencies \[]:
    -   @llm-ui/react@0.10.0
    -   @llm-ui/shared@0.10.0

## @llm-ui/markdown@0.10.0

### Patch Changes

-   Updated dependencies \[]:
    -   @llm-ui/react@0.10.0

## @llm-ui/shared@0.10.0

### Patch Changes

-   Updated dependencies \[]:
    -   @llm-ui/react@0.10.0

## @llm-ui/react@0.10.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>